Correct clean existing cache files during startup (#1237)
* Correct clean existing cache files during startup

Also improve warnings on immediate rejection from cache.

Fixes #1183.

* [CR] Use tfs.eviction member where possible

* [CR] Fix workspace cleanup on startup

* [CR] Keep deleting files when ristretto cache.Set() fails

A synchronous failure is due to the ristretto worker goroutine not being set up
correctly (possibly during cache shutdown), so onReject might not be called.

But continue to log nicely, of course, to debug such weird behaviour.

* Keep evidence of failed tier_fs tests

This is a temporary directory that is left behind on test failure.  In CI it does not matter
because the container dies anyway; on a developer machine it helps.

Also improve test error detection in TestStartup.

+++ pyramid/tier_fs_test.go
@@ -14,6 +14,7 @@ import (
 	"testing"
 
 	"github.com/google/uuid"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/treeverse/lakefs/block/mem"
 	"github.com/treeverse/lakefs/logging"
@@ -73,8 +74,12 @@ func TestStartup(t *testing.T) {
 	// cleanup
 	baseDir := path.Join(os.TempDir(), fsName)
 	defer func() {
+		if t.Failed() {
+			// Leave behind the evidence.
+			return
+		}
 		if err := os.RemoveAll(baseDir); err != nil {
-			t.Fatal("Remove all filed under", baseDir, err)
+			t.Fatal("Remove all files under", baseDir, err)
 		}
 	}()
 
@@ -112,16 +117,18 @@ func TestStartup(t *testing.T) {
 	}
 
 	dir, err := os.Open(workspacePath)
-	require.Nil(t, dir)
-	require.True(t, os.IsNotExist(err))
+	assert.Nil(t, dir, "expected to fail to open %s", workspacePath)
+	// os.IsNotExist does not look as hard to errors.Is; for errors returned directly from
+	// package os this does not matter.
+	assert.Error(t, err, os.ErrNotExist, "expected %s not to exist", workspacePath)
 
 	f, err := localFS.Open(ctx, namespace, filename)
 	defer func() { _ = f.Close() }()
-	require.NoError(t, err)
+	assert.NoError(t, err)
 
 	data, err := ioutil.ReadAll(f)
-	require.NoError(t, err)
-	require.Equal(t, content, data)
+	assert.NoError(t, err)
+	assert.Equal(t, content, data)
 }
 
 func testEviction(t *testing.T, namespaces ...string) {
